notesBackend/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cssselect==1.3.0
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.11
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
lxml_html_clean==0.4.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.3
packaging==25.0
pillow==12.0.0
preshed==3.0.10
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-file==3.0.0
rich==14.2.0
safetensors==0.6.2
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
